Update the value of noItemSelected
The isHidden properties of selectedLabelsView and thumbView are updated in the didSet callback of noItemSelected, but the value of noItemSelected isn't updated when an item is selected. Because of this, in the case when noItemSelected is set to false from an outside scope and then an item is selected, the value of noItemSelected would still remain false, although it should be true. This inconsistency could create bugs when the variable is inside the condition of an if statement.